+++ bin/cucumber-language-server.cjs
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
 require('source-map-support').install()
-const { startServer } = require('../dist/cjs/src/startServer')
-const { NodeParserAdapter } = require('@cucumber/language-service/node')
-startServer(new NodeParserAdapter())
+const { startNodeServer } = require('../dist/cjs/src/node/startNodeServer')
+startNodeServer()