+++ src/node/startNodeServer.ts
@@ -1,0 +1,7 @@
+import { NodeParserAdapter } from '@cucumber/language-service/node'
+
+import { startServer } from '../startServer'
+
+export function startNodeServer() {
+  startServer(new NodeParserAdapter())
+}